requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
requests==2.31.0
openai==1.12.0
gunicorn==21.2.0
SQLAlchemy==2.0.28
alembic==1.13.1
python-jose==3.3.0
